add: basic db setup/check with postgresql and render in menu
postgresql is hosted in render in free trial for test in this phase, currently only tested with the get method to retrieve the records of db. #note: the db was setup with create and insert query before hand using psycopg2 python library and some dummy record are present in the menu table of the db for now.